Exercise breakdown bugfix (#39)
Now when you click on a muscle group in the training volume graph, you
see a breakdown of this muscle group's volume per exercise type. The
design of this is very basic.

Also, a little fix: some Garmin exercises only have the category
present, not the name. This was already accounted for in the Strength
Workout list view, but not in the exercise or training volume views, so
fixed that here.

Should address #35