Use workout time when computing latest exercise date (#42)
The exercise start time can sometimes be not present. This ends up as an
Invalid Date object and that somehow gets promoted as the largest
(latest) date, which then can't be formatted properly for displaying and
causes errors. This should fix that as workout start time should always
be present.